Weather and Todo controllers fixed
