--- app/assets/javascripts/main.coffee.js
@@ -1,12 +1,0 @@
-var app = angular.module("Tasks", []);
-
-app.controller("TasksControl", function($scope, $http) {
-  $http.get('data/tasks.json').
-    success(function(data, status, headers, config) {
-      $scope.tasks = data;
-      console.log(data)
-    }).
-    error(function(data, status, headers, config) {
-      // log error
-    });
-});
