--- app/assets/javascripts/todoApp.js
@@ -1,7 +1,0 @@
-var app = angular.module('ToDoApp', ['ngDialog']);
-
-app.controller('MainCtrl', function ($scope, ngDialog) {
-    $scope.clickToOpen = function () {
-        ngDialog.open({ template: 'main/index.html' });
-    };
-});
