+++ app/views/layouts/angular.html.erb
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html ng-app="Todos">
 <head>
     <!-- Angular link -->
   <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.3.15/angular.min.js"></script>
